first refactoring of Map
